backend/pyproject.toml
[project]
name = "omnitrace-backend"
version = "0.1.0"
description = "Minimal backend test harness for OpenRouter + Weave"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "httpx>=0.28.0",
  "pydantic-settings>=2.10.0",
  "uvicorn[standard]>=0.35.0",
  "weave>=0.51.0"
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]
